adjusted some css content, reorganize cart file
