--- client/src/components/Navbar/NavbarStyles.scss
@@ -1,28 +1,0 @@
-.navbar-fixed{
-  position: fixed;
-  top: 0;
-  max-width: 1920px;
-  width: 100%; 
-  height: 70px;
-  border-bottom: 1px solid darkgray;
-
-  .navbar-container {
-    height: 60px;
-    // box-shadow: 0px 0px 20px .5px rgba(0, 0, 0, 0.566);
-    // border-radius: 5px;
-    opacity: .9;
-    .navbar-logo-place,
-    .navbar-option-place {
-      display: flex;
-      flex-direction: row;
-      align-items: center;
-      justify-content: center;
-      li {
-        margin-left: 35px;
-        margin-right: 35px;
-        font-size:20px;
-      }
-
-    }
-  }
-}
