+++ client/src/layouts/Header.js
@@ -1,24 +1,17 @@
 import { Link } from 'react-router-dom'
-import { useState } from 'react'
 import Logo from '../images/Logo/Logo'
-
+import { useState } from 'react'
 import './Header.scss'
 import { FaUserAlt } from 'react-icons/fa'
 import { BiShoppingBag } from 'react-icons/bi'
-
-
+import { ShoppingCart } from 'components/ShoppingCart/ShoppingCart'
 
 const Header = () => {
 
+  // for modals
   const [modalVisibility, setModalVisibility] = useState(false)
-
   const toggleModal = () => {
     setModalVisibility(!modalVisibility)
-  }
-  const closeModal = (event) => {
-    if (event.target === event.currentTarget) {
-      toggleModal()
-    }
   }
 
   return (
@@ -31,30 +24,30 @@ const Header = () => {
           ><Logo />
             <span className='title'><h1>Circular Journeys</h1></span>
           </Link>
-          <section>
-            <ul>
-              <li>
+          <section className='header-section'>
+            <ul className='header-ul'>
+              <li className='header-li'>
                 <Link to='/blog' title='Blog'>
-                  <span className='links'>部落格</span>
+                  <h5 className='links'>部落格</h5>
                 </Link>
               </li>
 
-              <li>
+              <li className='header-li'>
                 <Link to='/'>
-                  <span className='links'>自由行</span>
+                  <h5 className='links'>自由行</h5>
                 </Link>
               </li>
-              <li>
+              <li className='header-li'>
                 <Link to='/shop'>
-                  <span className='links'>商城</span>
+                  <h5 className='links'>商城</h5>
                 </Link>
               </li>
-              <li>
+              <li className='header-li'>
                 <button onClick={toggleModal}>
-                  <BiShoppingBag size={30} />
+                  <BiShoppingBag size={32} />
                 </button>
               </li>
-              <li>
+              <li className='header-li'>
                 <Link to='/login'>
                   <FaUserAlt size={30} />
                 </Link>
@@ -63,25 +56,15 @@ const Header = () => {
           </section>
         </div>
 
-        <div className='cart'>
-          {modalVisibility && (
-            <div className="modal-background" onClick={closeModal}>
-              <div className="modal-content">
-                <div className="close-button">
-                  <button onClick={toggleModal}>&times;</button>
-                </div>
+        {modalVisibility &&
+          <ShoppingCart
+            modalVisibility={modalVisibility}
+            setModalVisibility={setModalVisibility}
+            toggleModal={toggleModal}
+          />}
 
-                <h1>hi im modal</h1>
-                <h5 className="text-danger">
-                  <Link onClick={toggleModal} to="../checkout" title="結帳">結帳</Link>
-                </h5>
-
-                <button onClick={toggleModal}>Close</button>
-              </div>
-            </div>
-          )}
-        </div>
       </header>
+
     </>
   )
 }
